Replace double with float for atomic add

+++ llmc/gelu.cuh
@@ -56,17 +56,17 @@ void gelu_forward(floatX* out, const floatX* inp, int N, float* coord_check_data
     cudaCheck(cudaGetLastError());
     // data collection
     if (coord_check_data != NULL) {
-        double sum = 0.0;
-        double* sum_d;
-        cudaMalloc(&sum_d, sizeof(double));
-        cudaCheck(cudaMemsetAsync(sum_d, 0, sizeof(double), stream));
+        float sum = 0.0;
+        float* sum_d;
+        cudaMalloc(&sum_d, sizeof(float));
+        cudaCheck(cudaMemsetAsync(sum_d, 0, sizeof(float), stream));
         assert(N % WARP_SIZE == 0);
         int grid_size = CEIL_DIV(N, WARP_SIZE);
         abs_sum_kernel<<<grid_size, WARP_SIZE, 0, stream>>>(out, grid_size, WARP_SIZE, sum_d);
         cudaCheck(cudaGetLastError());
-        cudaCheck(cudaMemcpy(&sum, sum_d, sizeof(double), cudaMemcpyDeviceToHost));
+        cudaCheck(cudaMemcpy(&sum, sum_d, sizeof(float), cudaMemcpyDeviceToHost));
         cudaCheck(cudaFree(sum_d));
-        coord_check_data[cc_cnt] = (float)(sum / N);
+        coord_check_data[cc_cnt] = sum / N;
     }
 }
